[ES] Fixed variable type
Signed-óff-by: Jim Martens <github@2martens.de>

+++ es/blatt2/uebung2-1/uebung2-1.ino
@@ -10,7 +10,7 @@ int buttonTwoPin = 3;
 // used to achieve a 1kHz frequency
 // don't touch them
 int compareValue = 60;
-int rc = 10499;
+long rc = 10499;
 
 // this variable is affected by the buttonOne and buttonTwo actions
 int volatile counter = 0;
